eliminar y buscar productos

+++ src/app/modules/shared/components/confirm/confirm.component.ts
@@ -1,6 +1,7 @@
 import { Component, inject } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { CategoryService } from '../../services/category.service';
+import { ProductService } from '../../services/product.service';
 
 @Component({
   selector: 'app-confirm',
@@ -12,6 +13,7 @@ export class ConfirmComponent {
   private categoryService= inject(CategoryService);
   private dialogRef= inject(MatDialogRef);
   public data = inject(MAT_DIALOG_DATA);
+  private productService= inject(ProductService);
 
 
   onNoClick(){
@@ -20,16 +22,20 @@ export class ConfirmComponent {
 
   delete(){
     if (this.data != null){
-
-     
-      
+      if(this.data.module == "category"){}
         this.categoryService.deleteCategorie(this.data.id).
               subscribe( (data:any) =>{
                 this.dialogRef.close(1);
               }, (error: any) => {
                 this.dialogRef.close(2);
               })
-      
+            } else if (this.data.module == "product"){
+              this.productService.deleteProduct(this.data.id).
+              subscribe( (data:any) =>{
+                this.dialogRef.close(1);
+              }, (error: any) => {
+                this.dialogRef.close(2);
+              })
 
     } else {
       this.dialogRef.close(2);
